move AcalaDropAssets to common
